+++ examples/protocol-and-stream-muxing/3.js
@@ -37,7 +37,7 @@ const createNode = async () => {
       stream,
       async function (source) {
         for await (const msg of source) {
-          console.log(uint8ArrayToString(msg))
+          console.log(uint8ArrayToString(msg.subarray()))
         }
       }
     )
@@ -48,7 +48,7 @@ const createNode = async () => {
       stream,
       async function (source) {
         for await (const msg of source) {
-          console.log(uint8ArrayToString(msg))
+          console.log(uint8ArrayToString(msg.subarray()))
         }
       }
     )