+++ src/upgrader.ts
@@ -1,6 +1,6 @@
 import { logger } from '@libp2p/logger'
 import errCode from 'err-code'
-import { Dialer, Listener } from '@libp2p/multistream-select'
+import * as mss from '@libp2p/multistream-select'
 import { pipe } from 'it-pipe'
 // @ts-expect-error mutable-proxy does not export types
 import mutableProxy from 'mutable-proxy'
@@ -20,6 +20,7 @@ import type { Registrar } from '@libp2p/interface-registrar'
 import { DEFAULT_MAX_INBOUND_STREAMS, DEFAULT_MAX_OUTBOUND_STREAMS } from './registrar.js'
 import { TimeoutController } from 'timeout-abort-controller'
 import { abortableDuplex } from 'abortable-iterator'
+import { setMaxListeners } from 'events'
 
 const log = logger('libp2p:upgrader')
 
@@ -134,6 +135,11 @@ export class DefaultUpgrader extends EventEmitter<UpgraderEvents> implements Upg
     const timeoutController = new TimeoutController(this.inboundUpgradeTimeout)
 
     try {
+      // fails on node < 15.4
+      setMaxListeners?.(Infinity, timeoutController.signal)
+    } catch {}
+
+    try {
       const abortableStream = abortableDuplex(maConn, timeoutController.signal)
       maConn.source = abortableStream.source
       maConn.sink = abortableStream.sink
@@ -146,7 +152,7 @@ export class DefaultUpgrader extends EventEmitter<UpgraderEvents> implements Upg
         ({ setTarget: setPeer, proxy: proxyPeer } = mutableProxy())
         const idString = `${(Math.random() * 1e9).toString(36)}${Date.now()}`
         setPeer({ toString: () => idString })
-        maConn = metrics.trackStream({ stream: maConn, remotePeer: proxyPeer })
+        metrics.trackStream({ stream: maConn, remotePeer: proxyPeer })
       }
 
       log('starting the inbound connection upgrade')
@@ -247,7 +253,7 @@ export class DefaultUpgrader extends EventEmitter<UpgraderEvents> implements Upg
       ({ setTarget: setPeer, proxy: proxyPeer } = mutableProxy())
       const idString = `${(Math.random() * 1e9).toString(36)}${Date.now()}`
       setPeer({ toB58String: () => idString })
-      maConn = metrics.trackStream({ stream: maConn, remotePeer: proxyPeer })
+      metrics.trackStream({ stream: maConn, remotePeer: proxyPeer })
     }
 
     log('Starting the outbound connection upgrade')
@@ -345,9 +351,8 @@ export class DefaultUpgrader extends EventEmitter<UpgraderEvents> implements Upg
 
           void Promise.resolve()
             .then(async () => {
-              const mss = new Listener(muxedStream)
               const protocols = this.components.getRegistrar().getProtocols()
-              const { stream, protocol } = await mss.handle(protocols)
+              const { stream, protocol } = await mss.handle(muxedStream, protocols)
               log('%s: incoming stream opened on %s', direction, protocol)
 
               const metrics = this.components.getMetrics()
@@ -364,10 +369,16 @@ export class DefaultUpgrader extends EventEmitter<UpgraderEvents> implements Upg
               const streamCount = countStreams(protocol, 'inbound', connection)
 
               if (streamCount === incomingLimit) {
-                throw errCode(new Error('Too many incoming protocol streams'), codes.ERR_TOO_MANY_INBOUND_PROTOCOL_STREAMS)
+                muxedStream.abort(errCode(new Error(`Too many inbound protocol streams for protocol "${protocol}" - limit ${incomingLimit}`), codes.ERR_TOO_MANY_INBOUND_PROTOCOL_STREAMS))
+
+                return
               }
 
               muxedStream.stat.protocol = protocol
+
+              // If a protocol stream has been successfully negotiated and is to be passed to the application,
+              // the peerstore should ensure that the peer is registered with that protocol
+              this.components.getPeerStore().protoBook.add(remotePeer, [protocol]).catch(err => log.error(err))
 
               connection.addStream(muxedStream)
               this._onStream({ connection, stream: Object.assign(muxedStream, stream), protocol })
@@ -397,7 +408,6 @@ export class DefaultUpgrader extends EventEmitter<UpgraderEvents> implements Upg
 
         log('%s: starting new stream on %s', direction, protocols)
         const muxedStream = muxer.newStream()
-        const mss = new Dialer(muxedStream)
         const metrics = this.components.getMetrics()
         let controller: TimeoutController | undefined
 
@@ -407,33 +417,40 @@ export class DefaultUpgrader extends EventEmitter<UpgraderEvents> implements Upg
 
             controller = new TimeoutController(30000)
             options.signal = controller.signal
+
+            try {
+              // fails on node < 15.4
+              setMaxListeners?.(Infinity, controller.signal)
+            } catch {}
           }
 
-          let { stream, protocol } = await mss.select(protocols, options)
+          const { stream, protocol } = await mss.select(muxedStream, protocols, options)
 
           if (metrics != null) {
-            stream = metrics.trackStream({ stream, remotePeer, protocol })
+            metrics.trackStream({ stream, remotePeer, protocol })
           }
 
           const outgoingLimit = findOutgoingStreamLimit(protocol, this.components.getRegistrar())
           const streamCount = countStreams(protocol, 'outbound', connection)
 
           if (streamCount === outgoingLimit) {
-            throw errCode(new Error('Too many outgoing protocol streams'), codes.ERR_TOO_MANY_OUTBOUND_PROTOCOL_STREAMS)
+            const err = errCode(new Error(`Too many outbound protocol streams for protocol "${protocol}" - limit ${outgoingLimit}`), codes.ERR_TOO_MANY_OUTBOUND_PROTOCOL_STREAMS)
+            muxedStream.abort(err)
+
+            throw err
           }
 
+          // If a protocol stream has been successfully negotiated and is to be passed to the application,
+          // the peerstore should ensure that the peer is registered with that protocol
+          this.components.getPeerStore().protoBook.add(remotePeer, [protocol]).catch(err => log.error(err))
+
+          // after the handshake the returned stream can have early data so override
+          // the souce/sink
+          muxedStream.source = stream.source
+          muxedStream.sink = stream.sink
           muxedStream.stat.protocol = protocol
 
-          return Object.assign(
-            muxedStream,
-            stream,
-            {
-              stat: {
-                ...muxedStream.stat,
-                protocol
-              }
-            }
-          )
+          return muxedStream
         } catch (err: any) {
           log.error('could not create new stream', err)
 
@@ -531,12 +548,13 @@ export class DefaultUpgrader extends EventEmitter<UpgraderEvents> implements Upg
    * Attempts to encrypt the incoming `connection` with the provided `cryptos`
    */
   async _encryptInbound (connection: Duplex<Uint8Array>): Promise<CryptoResult> {
-    const mss = new Listener(connection)
     const protocols = Array.from(this.connectionEncryption.keys())
     log('handling inbound crypto protocol selection', protocols)
 
     try {
-      const { stream, protocol } = await mss.handle(protocols)
+      const { stream, protocol } = await mss.handle(connection, protocols, {
+        writeBytes: true
+      })
       const encrypter = this.connectionEncryption.get(protocol)
 
       if (encrypter == null) {
@@ -559,12 +577,13 @@ export class DefaultUpgrader extends EventEmitter<UpgraderEvents> implements Upg
    * The first `ConnectionEncrypter` module to succeed will be used
    */
   async _encryptOutbound (connection: MultiaddrConnection, remotePeerId: PeerId): Promise<CryptoResult> {
-    const mss = new Dialer(connection)
     const protocols = Array.from(this.connectionEncryption.keys())
     log('selecting outbound crypto protocol', protocols)
 
     try {
-      const { stream, protocol } = await mss.select(protocols)
+      const { stream, protocol } = await mss.select(connection, protocols, {
+        writeBytes: true
+      })
       const encrypter = this.connectionEncryption.get(protocol)
 
       if (encrypter == null) {
@@ -587,11 +606,12 @@ export class DefaultUpgrader extends EventEmitter<UpgraderEvents> implements Upg
    * muxer will be used for all future streams on the connection.
    */
   async _multiplexOutbound (connection: MultiaddrConnection, muxers: Map<string, StreamMuxerFactory>): Promise<{ stream: Duplex<Uint8Array>, muxerFactory?: StreamMuxerFactory}> {
-    const dialer = new Dialer(connection)
     const protocols = Array.from(muxers.keys())
     log('outbound selecting muxer %s', protocols)
     try {
-      const { stream, protocol } = await dialer.select(protocols)
+      const { stream, protocol } = await mss.select(connection, protocols, {
+        writeBytes: true
+      })
       log('%s selected as muxer protocol', protocol)
       const muxerFactory = muxers.get(protocol)
       return { stream, muxerFactory }
@@ -606,11 +626,12 @@ export class DefaultUpgrader extends EventEmitter<UpgraderEvents> implements Upg
    * selected muxer will be used for all future streams on the connection.
    */
   async _multiplexInbound (connection: MultiaddrConnection, muxers: Map<string, StreamMuxerFactory>): Promise<{ stream: Duplex<Uint8Array>, muxerFactory?: StreamMuxerFactory}> {
-    const listener = new Listener(connection)
     const protocols = Array.from(muxers.keys())
     log('inbound handling muxers %s', protocols)
     try {
-      const { stream, protocol } = await listener.handle(protocols)
+      const { stream, protocol } = await mss.handle(connection, protocols, {
+        writeBytes: true
+      })
       const muxerFactory = muxers.get(protocol)
       return { stream, muxerFactory }
     } catch (err: any) {